Add `--trust-remote-code` to `mlx-lm`
**Problems**

mlx-lm 0.26.3 added a new `--trust-remote-code` argument and without it,
it complains when loading models with a custom code.

Since mlx-lm dependencies is set to `>=0.25.2`, casual installation of
plamo-translate may end up with an error message saying
'Please pass the argument `trust_remote_code=True` to allow custom code
to be run.' that confuses users.

**Solution**

Add `--trust-remote-code` to the command line arguments of mlx-lm,
also increment the dependency version to `>=0.26.3`.